restructuring components
