+++ src/App.js
@@ -2,13 +2,13 @@ import React from 'react'
 import './App.css'
 import { Nav, Navbar, Button } from 'react-bootstrap'
 import { Route, Switch, Link, Redirect, useHistory } from 'react-router-dom'
-import Randomizer from './components/Randomizer'
-import AddForm from './components/AddForm'
+import Randomizer from './components/Randomizer/Randomizer'
+import AddForm from './components/Restaurants/AddForm/AddForm'
 import NotFound from './components/error/NotFound'
 import restaurantService from './services/restaurant'
 import categoryService from './services/category'
 
-import RestaurantList from './components/RestaurantList'
+import RestaurantList from './components/Restaurants/RestaurantList/RestaurantList'
 import LoginForm from './components/auth/LoginForm'
 import authService from './services/authentication'
 import CategoryForm from './components/Categories/CategoryForm/CategoryForm'
