+++ src/components/Restaurants/RestaurantEntry/RestaurantEntry.css
